+++ python/GafferImageTest/ImageSamplerTest.py
@@ -88,9 +88,7 @@ class ImageSamplerTest( GafferImageTest.ImageTestCase ) :
 
 						sampler["pixel"].setValue( imath.V2f( x, y ) )
 
-						c = sampler["color"].getValue()
-						for i in range( 4 ):
-							self.assertAlmostEqual( c[i], [ x, y, 0, 0 ][i], places = 3 )
+						self.assertEqualWithAbsError( sampler["color"].getValue(), imath.Color4f( x, y, 0, 0 ), 0.001 )
 						hashes.add( str( sampler["color"].hash() ) )
 						count += 1